Sped up Xen Orchestra upgrade process.
Most importantly, the xo-server/xo-web build will only happen if the new
xen-orchestra source code is different from the currently-installed
build.

Additionally, the xen-orchestra source code will now be kept in a
separate source directory. So when a new upgrade happens, we pull any
updates into that directory, and copy that updated source tree to the
new install directory. This way, we're not performing a full clone each
time we do an update.